refactor(fs): consolidate filesystem utilities into oso_dev_util_helper
- Move filesystem functions from oso_dev_util to oso_dev_util_helper
- Consolidate duplicate implementations of path searching and crate detection
- Remove redundant code from oso_dev_util/src/fs.rs
- Add comprehensive filesystem utilities to oso_dev_util_helper/src/fs.rs
- Move check_oso_kernel function from oso_proc_macro_logic to oso_dev_util_helper
- Improve code organization and reduce duplication across development utilities
